komorei

Module canvas

Structs and enums used for drawing on a canvas.

A lot of APIs are largely based on raqote.

Structs

Color

A color with red, green, blue, and alpha components.

Path

A 2D path.

PathBuilder

A builder for paths.

Point

A 2D point.

Rect

A rectangle.

StrokeStyle

The configuration options for drawing strokes.

Enums

FontWeight

A standard typeface weight.

LineCap

The endpoint style of a line.

LineJoin

The style of connected line joins.

PathOp

Type Aliases

Angle

Transform