komorei

Trait Source

pub trait Source {
    // Required methods
    fn new() -> Self;
    fn get_search_anime_list(
        &self,
        query: Option<String>,
        page: i32,
        filters: Vec<FilterValue>,
    ) -> Result<AnimePageResult>;
    fn get_anime_update(
        &self,
        anime: Anime,
        needs_details: bool,
        needs_chapters: bool,
    ) -> Result<Anime>;
    fn get_stream_list(
        &self,
        anime: Anime,
        episode: Episode,
    ) -> Result<Vec<StreamInfo>>;
    fn get_stream(
        &self,
        anime: Anime,
        episode: Episode,
        stream: StreamInfo,
    ) -> Result<StreamData>;
}

The required functions a Komorei source must implement.

Required Methods

fn new() -> Self

Called to initialize a source.

If a source requires any setup before other functions are called, it should happen here.

fn get_search_anime_list( &self, query: Option<String>, page: i32, filters: Vec<FilterValue>, ) -> Result<AnimePageResult>

Returns the anime for a search query with filters.

fn get_anime_update( &self, anime: Anime, needs_details: bool, needs_chapters: bool, ) -> Result<Anime>

Updates a given anime with new details and chapters (episodes), as requested.

fn get_stream_list( &self, anime: Anime, episode: Episode, ) -> Result<Vec<StreamInfo>>

Returns the playable streams (servers) for a given anime episode.

fn get_stream( &self, anime: Anime, episode: Episode, stream: StreamInfo, ) -> Result<StreamData>

Resolves the stream data (media url, headers, subtitles) for a given stream.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".